Clean up the smoke tests some more
