+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jca/metrics/DataSourceCfgMetricUnitTestCase.java
@@ -22,12 +22,10 @@
 
 package org.jboss.as.test.integration.jca.metrics;
 
-import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.test.integration.management.jca.DsMgmtTestBase;
-import org.jboss.as.test.smoke.modular.utils.ShrinkWrapUtils;
-import org.jboss.shrinkwrap.api.Archive;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -44,10 +42,9 @@ import static junit.framework.Assert.assertTrue;
 @RunAsClient
 public class DataSourceCfgMetricUnitTestCase  extends DsMgmtTestBase{
 
-	@Deployment
-    public static Archive<?> getDeployment() {
+	@BeforeClass
+    public static void before() {
     	setBaseAddress("data-source", "DS");
-        return ShrinkWrapUtils.createEmptyJavaArchive("dummy");
     }
 
     @Test
