+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/deployment/rar/tests/EarPackagedDeploymentTestCase.java
@@ -19,20 +19,29 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.smoke.deployment.rar.examples;
+package org.jboss.as.test.smoke.deployment.rar.tests;
+
+import java.util.List;
 
 import javax.annotation.Resource;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.as.arquillian.api.ServerSetup;
+import org.jboss.as.arquillian.container.ManagementClient;
+import org.jboss.as.connector.subsystems.resourceadapters.Namespace;
+import org.jboss.as.connector.subsystems.resourceadapters.ResourceAdaptersExtension.ResourceAdapterSubsystemParser;
+import org.jboss.as.test.integration.management.base.AbstractMgmtServerSetupTask;
 import org.jboss.as.test.integration.management.base.AbstractMgmtTestBase;
 import org.jboss.as.test.integration.management.base.ContainerResourceMgmtTestBase;
 import org.jboss.as.test.integration.management.util.MgmtOperationException;
+import org.jboss.as.test.shared.FileUtils;
 import org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject1;
-import org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject2;
 import org.jboss.as.test.smoke.deployment.rar.MultipleConnectionFactory1;
+import org.jboss.dmr.ModelNode;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
+import org.jboss.shrinkwrap.api.spec.EnterpriseArchive;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.jboss.shrinkwrap.api.spec.ResourceAdapterArchive;
 import org.jboss.staxmapper.XMLElementReader;
@@ -44,12 +53,32 @@ import static org.junit.Assert.assertNotNull;
 
 
 /**
- * @author <a href="vrastsel@redhat.com">Vladimir Rastseluev</a>
- *         JBQA-5736 -IronJacamar deployment test
+ * @author <a href="robert.reimann@googlemail.com">Robert Reimann</a>
+ *         Deployment of a RAR packaged inside an EAR.
  */
 @RunWith(Arquillian.class)
-public class IronJacamarActivationTestCase extends ContainerResourceMgmtTestBase {
+@ServerSetup(EarPackagedDeploymentTestCase.EarPackagedDeploymentTestCaseSetup.class)
+public class EarPackagedDeploymentTestCase extends ContainerResourceMgmtTestBase {
 
+    static class EarPackagedDeploymentTestCaseSetup extends AbstractMgmtServerSetupTask {
+
+        @Override
+        public void doSetup(final ManagementClient managementClient) throws Exception{
+                String xml = FileUtils.readFile(EarPackagedDeploymentTestCase.class, "ear_packaged.xml");
+                List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdapterSubsystemParser());
+                executeOperation( operationListToCompositeOperation(operations));
+        }
+
+        @Override
+        public void tearDown(final ManagementClient managementClient, final String containerId) throws Exception{
+
+            final ModelNode address = new ModelNode();
+            address.add("subsystem", "resource-adapters");
+            address.add("resource-adapter", "ear_packaged.ear#ear_packaged.rar");
+            address.protect();
+            remove(address);
+        }
+    }
 
     /**
      * Define the deployment
@@ -57,33 +86,36 @@ public class IronJacamarActivationTestCase extends ContainerResourceMgmtTestBase
      * @return The deployment archive
      */
     @Deployment
-    public static ResourceAdapterArchive createDeployment() throws Exception {
-        String deploymentName = "archive_ij.rar";
+    public static EnterpriseArchive createDeployment() throws Exception {
+
+        String deploymentName = "ear_packaged.ear";
+        String subDeploymentName = "ear_packaged.rar";
 
         ResourceAdapterArchive raa =
-                ShrinkWrap.create(ResourceAdapterArchive.class, deploymentName);
+                ShrinkWrap.create(ResourceAdapterArchive.class, subDeploymentName);
         JavaArchive ja = ShrinkWrap.create(JavaArchive.class, "multiple.jar");
         ja.addPackage(MultipleConnectionFactory1.class.getPackage()).
-                addClasses(IronJacamarActivationTestCase.class, MgmtOperationException.class, XMLElementReader.class, XMLElementWriter.class);
+                addClasses(EarPackagedDeploymentTestCase.class,  MgmtOperationException.class, XMLElementReader.class, XMLElementWriter.class);
 
         ja.addPackage(AbstractMgmtTestBase.class.getPackage());
         raa.addAsLibrary(ja);
 
-        raa.addAsManifestResource("rar/" + deploymentName + "/META-INF/ra.xml", "ra.xml")
-                .addAsManifestResource("rar/" + deploymentName + "/META-INF/ironjacamar.xml", "ironjacamar.xml")
+        raa.addAsManifestResource("rar/" + subDeploymentName + "/META-INF/ra.xml", "ra.xml")
                 .addAsManifestResource(new StringAsset("Dependencies: org.jboss.as.controller-client,org.jboss.dmr,org.jboss.as.cli\n"), "MANIFEST.MF");
 
-        return raa;
+        final EnterpriseArchive ear = ShrinkWrap.create(EnterpriseArchive.class, deploymentName);
+        ear.addAsModule(raa);
+        ear.addAsManifestResource("rar/" + deploymentName + "/META-INF/application.xml", "application.xml");
+        return ear;
     }
 
     @Resource(mappedName = "java:jboss/name1")
     private MultipleConnectionFactory1 connectionFactory1;
 
+
     @Resource(mappedName = "java:jboss/Name3")
     private MultipleAdminObject1 adminObject1;
 
-    @Resource(mappedName = "java:jboss/Name4")
-    private MultipleAdminObject2 adminObject2;
 
     /**
      * Test configuration
@@ -92,9 +124,8 @@ public class IronJacamarActivationTestCase extends ContainerResourceMgmtTestBase
      */
     @Test
     public void testConfiguration() throws Throwable {
+
         assertNotNull("CF1 not found", connectionFactory1);
         assertNotNull("AO1 not found", adminObject1);
-        assertNotNull("AO2 not found", adminObject2);
     }
-
 }
