+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/deployment/rar/tests/PureTestCase.java
@@ -19,26 +19,31 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.smoke.deployment.rar.examples;
+package org.jboss.as.test.smoke.deployment.rar.tests;
 
-import java.io.IOException;
 import java.util.List;
+import java.util.Set;
 
-import javax.annotation.Resource;
+import javax.inject.Inject;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.arquillian.container.ManagementClient;
+import org.jboss.as.connector.ConnectorServices;
 import org.jboss.as.connector.subsystems.resourceadapters.Namespace;
 import org.jboss.as.connector.subsystems.resourceadapters.ResourceAdaptersExtension.ResourceAdapterSubsystemParser;
 import org.jboss.as.test.integration.management.base.AbstractMgmtServerSetupTask;
 import org.jboss.as.test.integration.management.base.AbstractMgmtTestBase;
 import org.jboss.as.test.integration.management.base.ContainerResourceMgmtTestBase;
 import org.jboss.as.test.integration.management.util.MgmtOperationException;
-import org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject1;
-import org.jboss.as.test.smoke.deployment.rar.MultipleConnectionFactory1;
+import org.jboss.as.test.shared.FileUtils;
+import org.jboss.as.test.smoke.deployment.rar.inflow.PureInflowResourceAdapter;
 import org.jboss.dmr.ModelNode;
+import org.jboss.jca.core.spi.mdr.MetadataRepository;
+import org.jboss.jca.core.spi.rar.ResourceAdapterRepository;
+import org.jboss.msc.service.ServiceContainer;
+import org.jboss.msc.service.ServiceController;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
@@ -54,32 +59,32 @@ import static org.junit.Assert.assertNotNull;
 
 /**
  * @author <a href="vrastsel@redhat.com">Vladimir Rastseluev</a>
- *         JBQA-5740 multiple resources deployment
+ *         JBQA-5742 -Pure RA deployment test
  */
 @RunWith(Arquillian.class)
-@ServerSetup(MultipleActivationTestCase.MultipleActivationTestCaseSetup.class)
-public class MultipleActivationTestCase extends ContainerResourceMgmtTestBase {
+@ServerSetup(PureTestCase.PureTestCaseSetup.class)
+public class PureTestCase extends ContainerResourceMgmtTestBase {
 
-    static class MultipleActivationTestCaseSetup extends AbstractMgmtServerSetupTask {
+    static class PureTestCaseSetup extends AbstractMgmtServerSetupTask {
 
         @Override
         public void doSetup(final ManagementClient managementClient) throws Exception {
-                String xml = readXmlResource(System.getProperty("jbossas.ts.submodule.dir") + "/src/test/resources/config/simple.xml");
-                List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdapterSubsystemParser());
-                executeOperation(operationListToCompositeOperation(operations));
+            String xml = FileUtils.readFile(PureTestCase.class, "pure.xml");
+            List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdapterSubsystemParser());
+            executeOperation(operationListToCompositeOperation(operations));
+
         }
 
         @Override
-        public void tearDown(final ManagementClient managementClient, final String containerId) throws IOException, MgmtOperationException {
+        public void tearDown(final ManagementClient managementClient, final String containerId) throws Exception {
 
             final ModelNode address = new ModelNode();
             address.add("subsystem", "resource-adapters");
-            address.add("resource-adapter", "archive.rar");
+            address.add("resource-adapter", "pure.rar");
             address.protect();
             remove(address);
         }
     }
-
 
     /**
      * Define the deployment
@@ -88,34 +93,25 @@ public class MultipleActivationTestCase extends ContainerResourceMgmtTestBase {
      */
     @Deployment
     public static ResourceAdapterArchive createDeployment() throws Exception {
-        String deploymentName = "archive.rar";
+        String deploymentName = "pure.rar";
 
         ResourceAdapterArchive raa =
                 ShrinkWrap.create(ResourceAdapterArchive.class, deploymentName);
         JavaArchive ja = ShrinkWrap.create(JavaArchive.class, "multiple.jar");
-        ja.addPackage(MultipleConnectionFactory1.class.getPackage()).
-                addClasses(MultipleActivationTestCase.class, MgmtOperationException.class, XMLElementReader.class, XMLElementWriter.class);
-
+        ja.addClasses(PureInflowResourceAdapter.class, PureTestCase.class,
+                MgmtOperationException.class, XMLElementReader.class, XMLElementWriter.class, PureTestCaseSetup.class);
         ja.addPackage(AbstractMgmtTestBase.class.getPackage());
         raa.addAsLibrary(ja);
 
         raa.addAsManifestResource("rar/" + deploymentName + "/META-INF/ra.xml", "ra.xml")
-                .addAsManifestResource(new StringAsset("Dependencies: org.jboss.as.controller-client,org.jboss.dmr,org.jboss.as.cli\n"), "MANIFEST.MF");
-        ;
+                .addAsManifestResource(new StringAsset("Dependencies: org.jboss.as.controller-client,org.jboss.dmr,org.jboss.as.cli,javax.inject.api,org.jboss.as.connector\n"), "MANIFEST.MF");
+
         return raa;
     }
 
-    @Resource(mappedName = "java:jboss/name1")
-    private MultipleConnectionFactory1 connectionFactory1;
+    @Inject
+    public ServiceContainer serviceContainer;
 
-    @Resource(mappedName = "java:jboss/name2")
-    private MultipleConnectionFactory1 connectionFactory2;
-
-    @Resource(mappedName = "java:jboss/Name3")
-    private MultipleAdminObject1 adminObject1;
-
-    @Resource(mappedName = "java:jboss/Name4")
-    private MultipleAdminObject1 adminObject2;
 
     /**
      * Test configuration
@@ -123,12 +119,41 @@ public class MultipleActivationTestCase extends ContainerResourceMgmtTestBase {
      * @throws Throwable Thrown if case of an error
      */
     @Test
-    public void testConfiguration() throws Throwable {
-        assertNotNull("CF1 not found", connectionFactory1);
-        assertNotNull("CF2 not found", connectionFactory2);
-        assertNotNull("AO1 not found", adminObject1);
-        assertNotNull("AO2 not found", adminObject2);
-        assertEquals("not equal AOs", adminObject1, adminObject2);
+    public void testRegistryConfiguration() throws Throwable {
+        ServiceController<?> controller = serviceContainer.getService(ConnectorServices.RA_REPOSITORY_SERVICE);
+        assertNotNull(controller);
+        ResourceAdapterRepository repository = (ResourceAdapterRepository) controller.getValue();
+        assertNotNull(repository);
+        Set<String> ids = repository.getResourceAdapters();
+
+        assertNotNull(ids);
+        //On a running server it's 3 beacause HornetQResourceAdapter is always present  + ra itself and 1 actrivation from DMR
+        assertEquals(2, ids.size());
+
+        for (String piId : ids) {
+            assertNotNull(piId);
+            System.out.println("PID:" + piId);
+            assertNotNull(repository.getResourceAdapter(piId));
+        }
+
     }
 
+    @Test
+    public void testMetadataConfiguration() throws Throwable {
+        ServiceController<?> controller = serviceContainer.getService(ConnectorServices.IRONJACAMAR_MDR);
+        assertNotNull(controller);
+        MetadataRepository repository = (MetadataRepository) controller.getValue();
+        assertNotNull(repository);
+        Set<String> ids = repository.getResourceAdapters();
+
+        assertNotNull(ids);
+        //on a running server it's always 2 beacause HornetQResourceAdapter is always present
+        assertEquals(1, ids.size());
+
+        for (String piId : ids) {
+            assertNotNull(piId);
+            System.out.println("PID:" + piId);
+            assertNotNull(repository.getResourceAdapter(piId));
+        }
+    }
 }
