+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/deployment/rar/tests/RaTestConnectionTestCase.java
@@ -19,22 +19,20 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.smoke.deployment.rar.examples;
+package org.jboss.as.test.smoke.deployment.rar.tests;
 
 import java.util.List;
-import javax.naming.Context;
+
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.as.arquillian.api.ContainerResource;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.arquillian.container.ManagementClient;
-import org.jboss.as.connector.ConnectorServices;
 import org.jboss.as.connector.subsystems.resourceadapters.Namespace;
-import org.jboss.as.connector.subsystems.resourceadapters.ResourceAdaptersExtension;
+import org.jboss.as.connector.subsystems.resourceadapters.ResourceAdaptersExtension.ResourceAdapterSubsystemParser;
 import org.jboss.as.test.integration.management.base.AbstractMgmtServerSetupTask;
 import org.jboss.as.test.integration.management.base.ContainerResourceMgmtTestBase;
-import org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject1;
+import org.jboss.as.test.shared.FileUtils;
 import org.jboss.as.test.smoke.deployment.rar.MultipleConnectionFactory1;
 import org.jboss.dmr.ModelNode;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -45,87 +43,78 @@ import org.junit.runner.RunWith;
 
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP_ADDR;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 
 /**
  * @author <a href="vrastsel@redhat.com">Vladimir Rastseluev</a>
- *         JBQA-5841 test for RA activation
+ *        JBQA-5967 test connection in pool
  */
 @RunWith(Arquillian.class)
 @RunAsClient
-@ServerSetup(AfterResourceCreationDeploymentTestCase.AfterResourceCreationDeploymentTestCaseSetup.class)
-public class AfterResourceCreationDeploymentTestCase extends ContainerResourceMgmtTestBase {
+@ServerSetup(RaTestConnectionTestCase.RaTestConnectionTestCaseSetup.class)
+public class RaTestConnectionTestCase extends ContainerResourceMgmtTestBase {
+    private static ModelNode address;
+    private static String deploymentName = "testcon_mult.rar";
 
-    static String deploymentName = "basic-after.rar";
-
-    @ContainerResource
-    private Context context;
-
-
-    static class AfterResourceCreationDeploymentTestCaseSetup extends AbstractMgmtServerSetupTask {
+    static class RaTestConnectionTestCaseSetup extends AbstractMgmtServerSetupTask {
 
         @Override
-        public void tearDown(final ManagementClient managementClient, final String containerId) throws Exception{
-            final ModelNode address = new ModelNode();
+        public void doSetup(final ManagementClient managementClient) throws Exception {
+            address = new ModelNode();
             address.add("subsystem", "resource-adapters");
             address.add("resource-adapter", deploymentName);
             address.protect();
-            remove(address);
+            String xml = FileUtils.readFile(RaTestConnectionTestCase.class, "testcon_multiple.xml");
+            List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdapterSubsystemParser());
+            System.out.println(operations);
+            executeOperation(operationListToCompositeOperation(operations));
         }
 
         @Override
-        protected void doSetup(final ManagementClient managementClient) throws Exception {
-
+        public void tearDown(final ManagementClient managementClient, final String containerId) throws Exception {
+            remove(address);
         }
     }
 
-    private void setup() throws Exception {
-        String xml = readXmlResource(System.getProperty("jbossas.ts.submodule.dir") + "/src/test/resources/config/basic-after.xml");
-        List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdaptersExtension.ResourceAdapterSubsystemParser());
-        executeOperation(operationListToCompositeOperation(operations));
-
-        //since it is created after deployment it needs activation
-        final ModelNode address = new ModelNode();
-        address.add("subsystem", "resource-adapters");
-        address.add("resource-adapter", deploymentName);
-        address.protect();
-        final ModelNode operation = new ModelNode();
-        operation.get(OP).set("activate");
-        operation.get(OP_ADDR).set(address);
-        executeOperation(operation);
-
-    }
 
     /**
      * Define the deployment
      *
      * @return The deployment archive
      */
-    @Deployment(name = "basic-after.rar")
-    public static ResourceAdapterArchive createDeployment() throws Exception {
+   @Deployment
+    public static ResourceAdapterArchive createDeployment()  throws Exception{
 
 
         ResourceAdapterArchive raa =
                 ShrinkWrap.create(ResourceAdapterArchive.class, deploymentName);
-        JavaArchive ja = ShrinkWrap.create(JavaArchive.class, "multiple.jar");
+         JavaArchive ja = ShrinkWrap.create(JavaArchive.class,  "multiple.jar");
         ja.addPackage(MultipleConnectionFactory1.class.getPackage());
         raa.addAsLibrary(ja);
-
         raa.addAsManifestResource("rar/" + deploymentName + "/META-INF/ra.xml", "ra.xml");
         return raa;
     }
 
-    /**
-     * Test configuration
-     *
-     * @throws Throwable Thrown if case of an error
-     */
+
     @Test
-    public void testConfiguration() throws Throwable {
-        setup();
-        MultipleAdminObject1 adminObject1 = (MultipleAdminObject1) context.lookup("after/Name3");
-        assertNotNull("AO1 not found", adminObject1);
+    public void testConnection() throws Exception{
+        ModelNode testAddress=address.clone();
+        testAddress.add("connection-definitions","Pool1");
+        ModelNode op=new ModelNode();
+        op.get(OP).set("test-connection-in-pool");
+        op.get(OP_ADDR).set(testAddress);
+        assertTrue(executeOperation(op).asBoolean());
+
     }
 
+    @Test
+    public void flushConnections() throws Exception{
+        ModelNode testAddress=address.clone();
+        testAddress.add("connection-definitions","Pool1");
+        ModelNode op=new ModelNode();
+        op.get(OP).set("flush-idle-connection-in-pool");
+        op.get(OP_ADDR).set(testAddress);
+        executeOperation(op);
+    }
 }
