+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/mgmt/servermodule/marker.txt
