+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/web/WarTestCase.java
@@ -22,13 +22,14 @@
 package org.jboss.as.test.smoke.web;
 
 import java.net.URL;
+import java.util.concurrent.TimeUnit;
 
 import junit.framework.Assert;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.as.test.smoke.modular.utils.PollingUtils;
+import org.jboss.as.test.integration.common.HttpRequest;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
@@ -69,8 +70,6 @@ public class WarTestCase {
 
     private String performCall(String urlPattern, String param) throws Exception {
         URL url = new URL(this.url.toExternalForm() + urlPattern + "?input=" + param);
-        PollingUtils.UrlConnectionTask task = new PollingUtils.UrlConnectionTask(url);
-        PollingUtils.retryWithTimeout(10000, task);
-        return task.getResponse();
+        return HttpRequest.get(url.toExternalForm(), 10, TimeUnit.SECONDS);
     }
 }
